+++ include/livox_lidar_def.h
@@ -32,7 +32,7 @@
 #pragma pack(1)
 
 #define LIVOX_LIDAR_SDK_MAJOR_VERSION       1
-#define LIVOX_LIDAR_SDK_MINOR_VERSION       3
+#define LIVOX_LIDAR_SDK_MINOR_VERSION       4
 #define LIVOX_LIDAR_SDK_PATCH_VERSION       3
 
 #define kBroadcastCodeSize 16
@@ -67,6 +67,7 @@ typedef enum {
   kLivoxLidarTypeHAP = 15,
   kLivoxLidarTypePA = 16,
   kLivoxLidarTypeMid360s = 35,
+  kLivoxLidarTypeAvia2 = 40,
 } LivoxLidarDeviceType;
 
 typedef enum {
@@ -98,7 +99,12 @@ typedef enum {
   kKeyFusaEn                  = 0x001D,
   kKeyForceHeatEn             = 0x001E,
   kKeySetEscMode              = 0x0021,
+  kKeySetFovMode              = 0x0022,
+  kKeySetEchoMode             = 0x0024,
+  kKeySetNTPServerIp          = 0x0025,
   kKeySetPpsSyncMode          = 0x0026,
+  kKeySetITOCtrl              = 0x0027,
+  kKeySetFogNoiseFilter       = 0x0028,
   kKeySetImuRange             = 0x002B,
 
   kKeyLogParamSet             = 0x7FFF,
@@ -166,6 +172,19 @@ typedef struct {
 } LivoxLidarImuRawPoint;
 
 typedef struct {
+  int32_t x1;            /**< X axis, Unit:mm */
+  int32_t y1;            /**< Y axis, Unit:mm */
+  int32_t z1;            /**< Z axis, Unit:mm */
+  uint8_t reflectivity1; /**< Reflectivity */
+  uint8_t tag1;          /**< Tag */
+  int32_t x2;            /**< X axis, Unit:mm */
+  int32_t y2;            /**< Y axis, Unit:mm */
+  int32_t z2;            /**< Z axis, Unit:mm */
+  uint8_t reflectivity2; /**< Reflectivity */
+  uint8_t tag2;          /**< Tag */
+} LivoxLidarDoubleEchoRawPoint;
+
+typedef struct {
   int32_t x;            /**< X axis, Unit:mm */
   int32_t y;            /**< Y axis, Unit:mm */
   int32_t z;            /**< Z axis, Unit:mm */
@@ -193,7 +212,8 @@ typedef enum {
   kLivoxLidarImuData = 0,
   kLivoxLidarCartesianCoordinateHighData = 0x01,
   kLivoxLidarCartesianCoordinateLowData = 0x02,
-  kLivoxLidarSphericalCoordinateData = 0x03
+  kLivoxLidarSphericalCoordinateData = 0x03,
+  kLivoxLidarDoubleEchoData          = 0x11
 } LivoxLidarPointDataType;
 
 typedef enum {
@@ -264,10 +284,37 @@ typedef enum {
   kLivoxPpsSyncNormal = 0x00,
   kLivoxPpsSyncSpec = 0x01,
 } LivoxLidarPpsSyncMode;
+
+typedef enum {
+  kLivoxFogNoiseFilterDisable = 0x00,
+  kLivoxRainFilterMode= 0x01,
+  kLivoxFogFilterMode= 0x02
+} LivoxFogNoiseFilterMode;
+
+typedef enum {
+  kLivoxItoCtrlDisable = 0x00,
+  kLivoxItoCtrlEnable = 0x01,
+  kLivoxItoCtrlAuto   = 0x02
+} LivoxLidarItoCtrlMode;
+
 typedef enum {
   kLivoxEscSpeedNormal = 0x00,
   kLivoxEscSpeedSlow = 0x01,
 } LivoxLidarEscMode;
+
+typedef struct {
+  char host_ip[16];
+} NTPServerIpInfo;
+
+typedef enum {
+  kLivoxSmallFovMode = 0x00,
+  kLivoxBigFovMode   = 0x01,
+} LivoxLidarFovMode;
+
+typedef enum {
+  kLivoxStrongEchoMode = 0x00,
+  kLivoxFirstEchoMode  = 0x01,
+} LivoxLidarEchoMode;
 
 typedef enum {
   kLivoxLidarImuOutRate200Hz = 0x00,
@@ -422,6 +469,11 @@ typedef struct {
   uint8_t             imu_data_en;              // 0x001C
   uint8_t             fusa_en;                  // 0x001D
   uint8_t             esc_mode;                 // 0x0021
+  uint8_t             fov_mode;                 // 0x0022
+  uint8_t             echo_mode;                // 0x0024
+  NTPServerIpInfo     ntp_server_ip;            // 0x0025
+  uint8_t             ito_mode;                 // 0x0027
+  uint8_t             fog_noise_filter;         // 0x0028
 
   char                sn[16];                   // 0x8000
   char                product_info[64];         // 0x8001