+++ sdk_core/command_handler/general_command_handler.cpp
@@ -31,6 +31,7 @@
 #include "command_handler/hap_command_handler.h"
 #include "command_handler/mid360_command_handler.h"
 #include "command_handler/mid360s_command_handler.h"
+#include "command_handler/avia2_command_handler.h"
 #include "logger_handler/logger_manager.h"
 #include "debug_point_cloud_handler/debug_point_cloud_manager.h"
 #include "base/logging.h"
@@ -80,6 +81,9 @@ bool GeneralCommandHandler::Init(std::shared_ptr<std::vector<LivoxLidarCfg>>& cu
   }
   if (lidars_command_handler_.find(kLivoxLidarTypeMid360s) == lidars_command_handler_.end()) {
     lidars_command_handler_[kLivoxLidarTypeMid360s].reset(new Mid360sCommandHandler(device_manager_));
+  }
+  if (lidars_command_handler_.find(kLivoxLidarTypeAvia2) == lidars_command_handler_.end()) {
+    lidars_command_handler_[kLivoxLidarTypeAvia2].reset(new Avia2CommandHandler(device_manager_));
   }
   AddDetectedLidar(custom_lidars_cfg_ptr);
   return true;
@@ -290,6 +294,10 @@ void GeneralCommandHandler::CreateCommandHandler(const uint8_t dev_type) {
       if (!(lidars_command_handler_[dev_type]->Init(custom_lidars_cfg_map_))) {
         LOG_ERROR("General command handler init failed, the lidar of type:{} command init failed.", dev_type);
       }
+    } else if (dev_type == kLivoxLidarTypeAvia2) {
+      if (!(lidars_command_handler_[dev_type]->Init(custom_lidars_cfg_map_))) {
+        LOG_ERROR("General command handler init failed, the lidar of type:{} command init failed.", dev_type);
+      }
     }
     
     return;
@@ -313,6 +321,12 @@ void GeneralCommandHandler::CreateCommandHandler(const uint8_t dev_type) {
     } else if (dev_type == kLivoxLidarTypeMid360s) {
       std::shared_ptr<Mid360sCommandHandler> mid360s_command_handler_ptr(new Mid360sCommandHandler(device_manager_));
       lidars_command_handler_[dev_type] = mid360s_command_handler_ptr;        
+      if (!(lidars_command_handler_[dev_type]->Init(is_view_))) {
+        LOG_ERROR("General command handler init failed, the lidar of type:{} command init failed.", dev_type);
+      }
+    } else if (dev_type == kLivoxLidarTypeAvia2) {
+      std::shared_ptr<Avia2CommandHandler> avia2_command_handler_ptr(new Avia2CommandHandler(device_manager_));
+      lidars_command_handler_[dev_type] = avia2_command_handler_ptr;        
       if (!(lidars_command_handler_[dev_type]->Init(is_view_))) {
         LOG_ERROR("General command handler init failed, the lidar of type:{} command init failed.", dev_type);
       }
@@ -693,6 +707,30 @@ bool GeneralCommandHandler::GetQueryLidarInternalInfoKeys(const uint32_t handle,
         kKeyLidarDiagStatus,
         kKeyFwType,
         kKeyHmsCode
+      };
+      key_sets.swap(tmp_key_sets);
+      return true;
+    } else if (dev_type == kLivoxLidarTypeAvia2){
+      std::set<ParamKeyName> tmp_key_sets {
+        kKeyPclDataType,
+        kKeyLidarIpCfg,
+        kKeyStateInfoHostIpCfg,
+        kKeyLidarPointDataHostIpCfg,
+        kKeyLidarImuHostIpCfg,
+        kKeyWorkMode,
+        kKeyImuDataEn,
+        kKeySetFovMode,
+        kKeySetEchoMode,
+        kKeySetNTPServerIp,
+        kKeySetITOCtrl,
+        kKeySetFogNoiseFilter,
+        kKeySn,
+        kKeyVersionApp,
+        kKeyVersionLoader,
+        kKeyCurWorkState,
+        kKeyCoreTemp,
+        kKeyFwType,
+        kKeyHmsCode,
       };
       key_sets.swap(tmp_key_sets);
       return true;